server/requirements.txt
# ============================================================
# Mezzofy AI Assistant — Python Dependencies
# Generated from CONFIG.md — do not manually edit
# ============================================================

# Core Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
pyyaml==6.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# LLM
anthropic==0.18.1
openai==1.12.0                     # Used for Kimi (OpenAI-compatible API)
tiktoken==0.6.0

# Microsoft 365 / Graph API
azure-identity==1.15.0
msgraph-sdk==1.2.0
msal==1.26.0

# Task Queue & Scheduler
celery[redis]==5.3.6
redis==5.0.1
flower==2.0.1

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Document Generation
reportlab==4.0.7
weasyprint==61.0
python-pptx==0.6.23
python-docx==1.1.0
pandas==2.1.4
pypdf2==3.0.1
openpyxl==3.1.2

# Web / Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1

# Media Processing
Pillow==10.2.0
opencv-python-headless==4.9.0.80
moviepy==1.0.3
pydub==0.25.1
pytesseract==0.3.10
openai-whisper==20231117
noisereduce==3.0.0

# Email (fallback)
aiosmtplib==2.0.2
jinja2==3.1.3

# Utilities
python-magic==0.4.27
tqdm==4.66.1
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
